remove intermediary file io task
