+++ components/src/main/java/org/folio/rest/workflow/components/StreamingFileReadTask.java
@@ -3,7 +3,11 @@ package org.folio.rest.workflow.components;
 import javax.persistence.Entity;
 
 @Entity
-public class StreamingFileReadTask extends AbstractFileTask {
+public class StreamingFileReadTask extends Task {
+
+  String path;
+
+  String workflow;
 
   Long delay;
 
@@ -17,6 +21,22 @@ public class StreamingFileReadTask extends AbstractFileTask {
   public StreamingFileReadTask(String name) {
     this();
     setName(name);
+  }
+
+  public String getPath() {
+    return path;
+  }
+
+  public void setPath(String path) {
+    this.path = path;
+  }
+
+  public String getWorkflow() {
+    return workflow;
+  }
+
+  public void setWorkflow(String workflow) {
+    this.workflow = workflow;
   }
 
   public Long getDelay() {
