+++ components/src/main/java/org/folio/rest/workflow/components/StreamingFileWriteTask.java
@@ -3,7 +3,11 @@ package org.folio.rest.workflow.components;
 import javax.persistence.Entity;
 
 @Entity
-public class StreamingFileWriteTask extends AbstractFileTask {
+public class StreamingFileWriteTask extends Task {
+
+  String path;
+
+  String workflow;
 
   public StreamingFileWriteTask() {
     super();
@@ -14,6 +18,22 @@ public class StreamingFileWriteTask extends AbstractFileTask {
   public StreamingFileWriteTask(String name) {
     this();
     setName(name);
+  }
+
+  public String getPath() {
+    return path;
+  }
+
+  public void setPath(String path) {
+    this.path = path;
+  }
+
+  public String getWorkflow() {
+    return workflow;
+  }
+
+  public void setWorkflow(String workflow) {
+    this.workflow = workflow;
   }
 
 }
